Typing updates and finalized migration to query hooks.
