+++ packages/esm-patient-list-app/src/patientListData/api-remote.ts
@@ -1,5 +1,15 @@
-import { Location, OpenmrsResource, openmrsFetch } from '@openmrs/esm-framework';
-import { AddPatientData, PatientListFilter, PatientListMember, PatientListType } from './types';
+import { openmrsFetch } from '@openmrs/esm-framework';
+import {
+  AddPatientData,
+  CohortResponse,
+  NewCohortData,
+  OpenmrsCohort,
+  OpenmrsCohortMember,
+  OpenmrsCohortRef,
+  PatientListFilter,
+  PatientListMember,
+  PatientListType,
+} from './types';
 
 const cohortUrl = '/ws/rest/v1/cohortm';
 
@@ -18,43 +28,6 @@ async function postData(url: string, data = {}, ac = new AbortController()) {
     body: JSON.stringify(data),
   });
   return response.data;
-}
-
-export interface OpenmrsCohort {
-  uuid: string;
-  resourceVersion: string;
-  name: string;
-  description: string;
-  attributes: Array<any>;
-  links: Array<any>;
-  location: Location | null;
-  groupCohort: boolean | null;
-  startDate: string | null;
-  endDate: string | null;
-  voidReason: string | null;
-  voided: boolean;
-  isStarred?: boolean;
-  type?: string;
-}
-
-export interface OpenmrsCohortRef {
-  cohort: OpenmrsCohortMember;
-}
-
-export interface OpenmrsCohortMember {
-  attributes: Array<any>;
-  description: string;
-  endDate: string;
-  name: string;
-  uuid: string;
-  patient: {
-    uuid: string;
-  };
-}
-
-interface CohortRepsonse<T> {
-  results: Array<T>;
-  error: any;
 }
 
 export async function getAllPatientLists(filter: PatientListFilter = {}, ac = new AbortController()) {
@@ -77,7 +50,7 @@ export async function getAllPatientLists(filter: PatientListFilter = {}, ac = ne
   const params = query.map(([key, value]) => `${key}=${encodeURIComponent(value)}`).join('&');
   const {
     data: { results, error },
-  } = await openmrsFetch<CohortRepsonse<OpenmrsCohort>>(`${cohortUrl}/cohort?${params}`, {
+  } = await openmrsFetch<CohortResponse<OpenmrsCohort>>(`${cohortUrl}/cohort?${params}`, {
     signal: ac.signal,
   });
 
@@ -91,7 +64,7 @@ export async function getAllPatientLists(filter: PatientListFilter = {}, ac = ne
 export async function getPatientListMembers(cohortUuid: string, ac = new AbortController()) {
   const {
     data: { results, error },
-  } = await openmrsFetch<CohortRepsonse<OpenmrsCohortMember>>(
+  } = await openmrsFetch<CohortResponse<OpenmrsCohortMember>>(
     `${cohortUrl}/cohortmember?cohort=${cohortUuid}&v=default`,
     {
       signal: ac.signal,
@@ -115,7 +88,7 @@ export async function getPatientListMembers(cohortUuid: string, ac = new AbortCo
 export async function getPatientListIdsForPatient(patientUuid: string, ac = new AbortController()) {
   const {
     data: { results, error },
-  } = await openmrsFetch<CohortRepsonse<OpenmrsCohortRef>>(
+  } = await openmrsFetch<CohortResponse<OpenmrsCohortRef>>(
     `${cohortUrl}/cohortmember?patient=${patientUuid}&v=default`,
     {
       signal: ac.signal,
@@ -131,11 +104,6 @@ export async function getPatientListIdsForPatient(patientUuid: string, ac = new 
 
 export async function addPatientToList(data: AddPatientData, ac = new AbortController()) {
   return postData(`${cohortUrl}/cohortmember`, data, ac);
-}
-
-export interface NewCohortData {
-  name: string;
-  description: string;
 }
 
 export async function createPatientList(cohort: NewCohortData, ac = new AbortController()) {
