+++ packages/esm-patient-list-app/src/patientListData/index.ts
@@ -2,3 +2,4 @@ export * from './api-remote';
 export * from './api-local';
 export * from './api';
 export * from './queries';
+export * from './types';
