+++ packages/esm-patient-list-app/src/patientListData/types.ts
@@ -33,12 +33,50 @@ export interface PatientListMember {
   id: string;
 }
 
-export interface PatientListMemberFilter {
-  // TODO
-}
-
 export interface AddPatientData {
   patient: string;
   cohort: string;
   startDate: string;
+}
+
+export interface OpenmrsCohort {
+  uuid: string;
+  resourceVersion: string;
+  name: string;
+  description: string;
+  attributes: Array<any>;
+  links: Array<any>;
+  location: Location | null;
+  groupCohort: boolean | null;
+  startDate: string | null;
+  endDate: string | null;
+  voidReason: string | null;
+  voided: boolean;
+  isStarred?: boolean;
+  type?: string;
+}
+
+export interface OpenmrsCohortRef {
+  cohort: OpenmrsCohortMember;
+}
+
+export interface OpenmrsCohortMember {
+  attributes: Array<any>;
+  description: string;
+  endDate: string;
+  name: string;
+  uuid: string;
+  patient: {
+    uuid: string;
+  };
+}
+
+export interface CohortResponse<T> {
+  results: Array<T>;
+  error: any;
+}
+
+export interface NewCohortData {
+  name: string;
+  description: string;
 }
